continue client strategy integration
