+++ connect/net_http.go
@@ -9,6 +9,14 @@ package connect
 
 
 
+// FIXME use netip.Addr instead of net.IP
+// FIXME use netip.Prefix instead of net.IPNet
+
+
+type DialTlsContextFunction = func(ctx context.Context, network string, addr string) (net.Conn, error)
+
+
+/*
 const defaultHttpTimeout = 60 * time.Second
 const defaultHttpConnectTimeout = 5 * time.Second
 const defaultHttpTlsTimeout = 5 * time.Second
@@ -29,7 +37,7 @@ func defaultClient() *http.Client {
 		Timeout: defaultHttpTimeout,
 	}
 }
-
+*/
 
 
 
@@ -89,7 +97,9 @@ type ClientStrategySettings struct {
 	// EnableUdp bool
 
 
-	LockInMinimumWeight float
+	// for gets and ws connects
+	ParallelBlockSize int
+
 
 
 	// non-extender minimum weight
@@ -100,47 +110,59 @@ type ClientStrategySettings struct {
 
 	// transportMode can be Tcp or Quic
 	// extender profile is a transport mode, server name, and port 
-	ExtenderProfiles []*ExtenderProfiles
 	ExpandExtenderProfileCount int
 
 	ExtenderNetworks []net.IpNet
 	// these are evaluated with DoH to grow the extender ips
 	ExtenderHostnames []string
-	InitialExtenderIpCount int
 	ExpandExtenderIpCount int
 
-	ExpandExtenderCount int
-	
-	ParallelBlockSize int
+	MaxExtenderCount int
 
 	// success / (success + error)
-	MinimumExtenderWeight float
-	MinimumExtenderWeightCount int
+	ExtenderMinimumWeight float
 
 	// drop dialers that have not had a successful connect in this timeout
 	ExtenderDropTimeout time.Duration
 
 
-	DefaultHttpTimeout time.Duration
-	DefaultHttpConnectTimeout time.Duration
-	DefaultHttpTlsTimeout time.Duration
+
+
+
+
+	TlsConfig *tls.Config
+
+	HttpTimeout time.Duration
+	HttpConnectTimeout time.Duration
+	HttpTlsTimeout time.Duration
 	
+
+	DohSettings *DohSettings
 	
+}
+
+func DefaultClientStrategySettings() *ClientStrategySettings {
+
 }
 
 
 // non-extender dialers are never dropped
 type clientDialer struct {
-	Dialer
-	ExtenderProfile
-
-	dialer DialContextFunc
+	dialTlsContext DialTlsContextFunction
 	extenderProfile ExtenderProfile
+	extenderIp net.IP
+	extenderSecret string
 
+
+	// these are locked under the client stategy mutex
 	successCount int
 	errorCount int
 	lastSuccessTime time.Time
 	lastErrorTime time.Time
+}
+
+func IsLastSuccess() bool {
+	dialer.lastSuccessTime < dialer.lastErrorTime
 }
 
 
@@ -150,6 +172,7 @@ type ClientStrategy struct {
 
 	// handhake timeout
 	// connect timeout
+	ctx context.Context
 
 
 	settings *ClientStrategySettings
@@ -157,7 +180,10 @@ type ClientStrategy struct {
 
 	mutex sync.Mutex
 	// dialers are only updated inside the mutex
-	dialers []*clientDialer
+	dialers map[*clientDialer]bool
+	resolvedExtenderIps []net.Ip
+
+	extenderIpSecrets map[net.IP]string
 
 
 	// FIXME dialers
@@ -176,43 +202,59 @@ type ClientStrategy struct {
 
 
 func DefaultClientStrategy() *ClientStrategy {
-	
+	return NewClientStrategy(DefaultClientStrategySettings())
 }
 
 
 // extender udp 53 to platform extender
-func NewClientStrategy() *ClientStrategy {
-	if len(extenderConfigs) == 0 {
-		// normal, resilient, extender udp 53 to platform extenders (extender.bringyour.com)
-	} else {
-		// use extender configs as given
-		// typically the system will use the following defaults
-		// extender tls 443
-		// extender quic 443
-		// extender udp 53
+func NewClientStrategy(ctx context.Context, settings *ClientStrategySettings) *ClientStrategy {
+	
+	// create dialers to match settings
+	dialers := []*clientDialer{}
+	resolvedExtenderIps := []net.Ip{}
+
+	if settings.EnableNormal {
+		// FIXME normal dial context
+	}
+	if settings.EnableResilient {
+		// FIXME tls frag, retransmit, tls frag + retransmit
 	}
 
+
+
+	return &ClientStrategy{
+		ctx, 
+		settings: settings,
+		dialers: dialers,
+		resolvedExtenderIps,
+	}
+
+
+}
+
+
+func (self *ClientStrategy) SetCustomExtenders(extenderIpSecrets map[net.IP]string) {
+	self.mutex.Lock()
+	defer self.mutex.Unlock()
+
+	self.extenderIpSecrets = Copy(extenderIpSecrets)
+}
+
+func (self *ClientStrategy) CustomExtenders() map[net.IP]string {
+	self.mutex.Lock()
+	defer self.mutex.Unlock()
+
+	return Copy(self.extenderIpSecrets)
 }
 
 
 
 // the normal strategy is not mixed with extenders
 
-// look up dns to find new extender ips
-func (self *ClientStrategy) expand() []*clientDialer {
-
-
-	// expand ips
-	// expand profiles
-	// now randomly expand the strategy pool by N by randomly pairing available ips with available profiles
-
-
-	// drop strategies that have not had a successful connect in DropTimeout
-}
 
 // FIXME rankDialers
 
-func (self *ClientStrategy) rankedDialers() map[*clientDialer]float32 {
+func (self *ClientStrategy) getDialerWeights() map[*clientDialer]float32 {
 
 }
 
@@ -230,8 +272,106 @@ func (self *ClientStrategy) Get(ctx context.Context, request *net.Request) (*net
 	// in blocks of ParallelBlockSize, run dialers in parallel
 	// when one finishes with success or error, try the next, until one succeeds
 	// if no more to try, call expand() and keep going with the parallel evaluation
+	// drop strategies that have not had a successful connect in DropTimeout
 
 
+
+	handleCtx, cancel := context.WithTimeout(ctx, self.settings.DefaultHttpTimeout)
+	defer cancel()
+
+
+
+
+	self.collapseExtenderDialers()
+
+
+	dialerWeights := self.getDialerWeights()
+	// descending
+	orderedDialers := Sort(maps.Keys(dialerWeights))
+
+
+	parallelDialers := []*clientDialer{}
+
+	for i := 0; i < len(orderedDialers) && orderedDialers[i].IsLastSuccess(); i += 1 {
+		RUN(handleCtx, orderedDialers[i])
+		if err == nil {
+			// FIXME return
+		} else {
+			parallelDialers = append(parallelDialers, orderedDialers[i])
+		}
+	}
+
+	WeightedShuffle(parallelDialers, dialerWeights)
+
+	j := 0
+	for j < len(parallelDialers) && j < ParallelBlockSize; j += 1 {
+		go RUN(handleCtx, paralleDialers[j])
+	}
+
+	for j < len(parallelDialers) {
+		select {
+		case <- self.ctx.Done():
+			// FIXME return
+		case <- ctx.Done():
+			// FIXME return
+		case err := <- result:
+			if err == nil {
+				// FIXME return
+			} else {
+				go RUN(handleCtx, paralleDialers[j])
+			}
+		}
+	}
+
+
+	// FIXME keep trying as long as there is time left
+
+	for {
+		select {
+		case <- self.ctx.Done():
+			// FIXME return
+		case <- handleCtx.Done():
+			// FIXME return
+		default:
+		}
+
+		self.collapseExtenderDialers()
+		expandedDialers, _ := self.expandExtenderDialers()
+		if 0 < len(expandedDialers) {
+			j = 0
+			for j < len(expandedDialers) && j < ParallelBlockSize; j += 1 {
+				go RUN(handleCtx, expandedDialers[j])
+			}
+			for j < len(expandedDialers) {
+				select {
+				case <- self.ctx.Done():
+					// FIXME return
+				case <- handleCtx.Done():
+					// FIXME return
+				case err := <- result:
+					if err == nil {
+						// FIXME return
+					} else {
+						go RUN(handleCtx, expandedDialers[j])
+					}
+				}
+			}
+
+		}
+	}
+
+
+
+
+
+
+
+
+
+
+
+
+	
 
 }
 
@@ -245,6 +385,64 @@ func (self *ClientStrategy) Post(ctx context.Context, request *net.Request) (*ne
 	// weighted sort all dialers
 	// sequentially try one at a time with success or error, try the next until one succeeds
 	// if no more to try, call expand() and keep going
+	// drop strategies that have not had a successful connect in DropTimeout
+
+	handleCtx, cancel := context.WithTimeout(ctx, self.settings.DefaultHttpTimeout)
+	defer cancel()
+
+
+	self.collapseExtenderDialers()
+
+
+
+	dialerWeights := self.getDialerWeights()
+	// descending
+	orderedDialers := Sort(maps.Keys(dialerWeights))
+
+	for i := 0; i < len(orderedDialers) && orderedDialers[i].IsLastSuccess(); i += 1 {
+		RUN(handleCtx, paralleDialers[j])
+		if err == nil {
+			// FIXME return
+		}
+	}
+	for i := 0; i < len(orderedDialers) && !orderedDialers[i].IsLastSuccess(); i += 1 {
+		RUN(handleCtx, paralleDialers[j])
+		if err == nil {
+			// FIXME return
+		}
+	}
+
+	for {
+		select {
+		case <- self.ctx.Done():
+			// FIXME return
+		case <- handleCtx.Done():
+			// FIXME return
+		default:
+		}
+		
+		self.collapseExtenderDialers()
+		expandedDialers, _ := self.expandExtenderDialers()
+		if 0 < len(expandedDialers) {
+			for i := 0; i < len(expandedDialers); i += 1 {
+				select {
+				case <- self.ctx.Done():
+					// FIXME return
+				case <- handleCtx.Done():
+					// FIXME return
+				default:
+				}
+
+				RUN(handleCtx, expandedDialers[j])
+				if err == nil {
+					// FIXME return
+				}
+
+					
+			}
+		}
+	}
+
 }
 
 
@@ -256,10 +454,358 @@ func (self *ClientStrategy) WsDialContext(ctx context.Context, url string, reque
 	// in blocks of ParallelBlockSize, run dialers in parallel
 	// when one finishes with success or error, try the next, until one succeeds
 	// if no more to try, call expand() and keep going with the parallel evaluation
+	// drop strategies that have not had a successful connect in DropTimeout
 
 	// success is websocket connects
 
+	/*
+	     for {
+-        wsDialer := &websocket.Dialer{
+-            // NetDialTLSContext: NewResilientTlsDialContext(self.dialer),
+-            NetDialContext: NewExtenderDialContext(ExtenderConnectModeQuic, self.dialer, TestExtenderConfig()),
+-            HandshakeTimeout: self.settings.WsHandshakeTimeout,
+-        }
+-
+         ws, err := func()(*websocket.Conn, error) {
+-            ws, _, err := wsDialer.DialContext(self.ctx, self.platformUrl, nil)
++            ws, _, err := self.clientStrategy.WsDialContext(self.ctx, self.platformUrl, nil)
+             if err != nil {
+                 return nil, err
+             }
+             */
+
+
+    // FIXME close all connections that were created except the first successful one
+
+
+	handleCtx, cancel := context.WithTimeout(ctx, self.settings.DefaultHttpTimeout)
+	defer cancel()
+
+
+
+
+	self.collapseExtenderDialers()
+
+
+	dialerWeights := self.getDialerWeights()
+	// descending
+	orderedDialers := Sort(maps.Keys(dialerWeights))
+
+	parallelDialers := []*clientDialer{}
+
+	for i := 0; i < len(orderedDialers) && orderedDialers[i].IsLastSuccess(); i += 1 {
+		RUN(handleCtx, orderedDialers[i])
+		if err == nil {
+			// FIXME return
+		} else {
+			parallelDialers = append(parallelDialers, orderedDialers[i])
+		}
+	}
+	WeightedShuffle(parallelDialers, dialerWeights)
+
+	j := 0
+	for j < len(parallelDialers) && j < ParallelBlockSize; j += 1 {
+		go RUN(handleCtx, paralleDialers[j])
+	}
+
+	for j < len(parallelDialers) {
+		select {
+		case <- self.ctx.Done():
+			// FIXME return
+		case <- ctx.Done():
+			// FIXME return
+		case err := <- result:
+			if err == nil {
+				// FIXME return
+			} else {
+				go RUN(handleCtx, paralleDialers[j])
+			}
+		}
+	}
+
+
+	// FIXME keep trying as long as there is time left
+
+	for {
+		select {
+		case <- self.ctx.Done():
+			// FIXME return
+		case <- handleCtx.Done():
+			// FIXME return
+		default:
+		}
+
+		self.collapseExtenderDialers()
+		expandedDialers, _ := self.expandExtenderDialers()
+		if 0 < len(expandedDialers) {
+			j = 0
+			for j < len(expandedDialers) && j < ParallelBlockSize; j += 1 {
+				go RUN(handleCtx, expandedDialers[j])
+			}
+			for j < len(expandedDialers) {
+				select {
+				case <- self.ctx.Done():
+					// FIXME return
+				case <- handleCtx.Done():
+					// FIXME return
+				case err := <- result:
+					if err == nil {
+						// FIXME return
+					} else {
+						go RUN(handleCtx, expandedDialers[j])
+					}
+				}
+			}
+
+		}
+	}
+
 }
+
+
+func (self *ClientStrategy) collapseExtenderDialers() {
+	mutex.Lock()
+	defer mutex.Unlock()
+
+	for dialer, _ := range self.dialers {
+		if dialer.IsExtender() && dialer.IsLastSuccess() {
+			if self.settings.ExtenderDropTimeout <= time.Now().Sub(dialer.lastErrorTime) {
+				delete(self.dialers, dialers)
+			}
+		}
+	}
+}
+
+
+// look up dns to find new extender ips
+func (self *ClientStrategy) expandExtenderDialers() ([]*clientDialer, []net.IP) {
+	mutex.Lock()
+	defer mutex.Unlock()
+
+	if settings.ExpandExtenderProfileCount <= 0 {
+		return []*clientDialer{}, []net.Id{}
+	}
+
+
+	// find current extender profiles
+	// find current ips
+
+	// expand ips
+	// expand profiles
+	// now randomly expand the strategy pool by N by randomly pairing available ips with available profiles
+
+
+	// FIXME enumerate the dialers and pull these out
+	visitedExtenderProfiles := map[ExtenderProfile]bool{}
+
+
+	visitedExtenderIps := []net.Id{}
+	visitedExtenderIpv4s := map[[4]byte]bool{}
+	visitedExtenderIpv6s := map[[16]byte]bool{}
+
+	for _, dialer := self.dialers {
+		if dialier.IsExtender() {
+			visitedExtenderProfiles[dialer.extenderProfile] = true
+			ip := dialer.extenderIp
+			if ip.IsIpv6() {
+				ipv6 := ip.ToIpv6()
+				visitedExtenderIpv6s[ipv6] = true
+			} else {
+				ipv4 := ip.ToIpv4()
+				visitedExtenderIpv4s[ipv4] = true
+			}
+		}
+	}
+
+
+
+	if settings.MaxExtenderCount <= len(visitedExtenderProfiles) {
+		// at maximum extenders
+		return []*clientDialer{}, []net.Id{}
+	}
+
+
+	deviceIpv4 := IsIpv4()
+	deviceIpv6 := IsIpv6()
+
+
+
+
+	extenderProfiles := EnumerateExtenderProfiles(min(settings.ExpandExtenderProfileCount, settings.MaxExtenderCount - len(visitedExtenderProfiles)), visitedExtenderProfiles)
+
+
+	extenderConfigs := []*ExtenderConfig{}
+
+	if len(self.extenderIpSecrets) == 0 {
+
+		// filter resolvedExtenderIps by visited
+		unusedExtenderIps := []net.Id{}
+		for _, ip := range self.resolvedExtenderIps {
+			if ip.IsIpv6() {
+				ipv6 := ip.ToIpv6()
+				if !visitedExtenderIpv6s[ipv6] {
+					unusedExtenderIps = append(unusedExtenderIps, ip)
+				}
+			} else {
+				ipv4 := ip.ToIpv4()
+				if !visitedExtenderIpv4s[ipv4] {
+					unusedExtenderIps = append(unusedExtenderIps, ip)
+				}
+			}
+		}
+
+		// expand the ips to have one new ip per profile
+		expandedExtenderIps := []net.IP{}
+		if len(unusedExtenderIps < len(extenderProfiles)) {
+
+
+			// iterate these for ips not used
+			for _, network := range settings.ExtenderNetworks {
+				if settings.ExpandExtenderIpCount <= len(expandedExtenderIps) {
+					break
+				}
+
+				if network.IP.IsIp4() && IsIpv4() {
+					_, bits := network.Mask.Bits()
+					prefix := netip.PrefixFrom(network.IP, bits)
+					for addr := prefix.Addr(); prefix.Contains(addr); addr = addr.Next() {
+						ipv6 := addr.As4()
+						if !visitedExtenderIpv4s[ipv4] {
+							visitedExtenderIpv4s[ipv4] = true
+							expandedExtenderIps = append(expandedExtenderIps, ip)
+						}
+					}
+				}
+				if network.IP.IsIp6() && IsIpv6() {
+
+					_, bits := network.Mask.Bits()
+					prefix := netip.PrefixFrom(network.IP, bits)
+					for addr := prefix.Addr(); prefix.Contains(addr); addr = addr.Next() {
+						ipv6 := addr.As16()
+						if !visitedExtenderIpv6s[ipv6] {
+							visitedExtenderIpv6s[ipv6] = true
+							expandedExtenderIps = append(expandedExtenderIps, ip)
+						}
+					}
+
+				}
+				
+
+			}
+
+			// if not enough ips, use DoH to load ips for the extender hostnames
+			if len(expandedExtenderIps) < settings.ExpandExtenderIpCount && 0 < len(settings.ExtenderHostnames) {
+
+				loadDoh(result map[string]map[string][]net.IP) {
+					for _, serverResult := range result {
+						for _, ips := range serverResult {
+							for _, ip := range ips {
+								if ip.IPv6() {
+									ipv6 := ip.ToIpv6()
+									if !visitedExtenderIpv6s[ipv6] {
+										visitedExtenderIpv6s[ipv6] = true
+										expandedExtenderIps = append(expandedExtenderIps, ip)
+									}
+								} else {
+									ipv4 := ip.ToIpv4()
+									if !visitedExtenderIpv4s[ipv4] {
+										visitedExtenderIpv4s[ipv4] = true
+										expandedExtenderIps = append(expandedExtenderIps, ip)
+									}
+								}
+							}
+						}
+					}
+				}
+
+				// the network can be both ipv4 and ipv6
+				if IsIpv4() {
+					result, err := DohQuery(ctx, 4, settings.ExtenderHostnames, "A")
+					if err == nil {
+						loadDoh(result)	
+					}
+				}
+				if IsIpv6() {
+					result, err := DohQuery(ctx, 6, settings.ExtenderHostnames, "AAAA")
+					if err == nil {
+						loadDoh(result)	
+					}
+				}
+				
+
+
+				
+			}
+
+			unusedExtenderIps = append(unusedExtenderIps, expandedExtenderIps...)
+		}
+
+		
+		// use unused ips first
+		Shuffle(unusedExtenderIps)
+		n := min(len(extenderProfiles), len(unusedExtenderIps))
+		for i := range n {
+			extenderConfig := &ExtenderConfig{
+				ExtenderProfile: extenderProfiles[i],
+				Ip: unusedExtenderIps[i],
+			}
+			extenderConfigs = append(extenderConfigs, extenderConfig)
+		}
+		if n < len(extenderProfiles) {
+			for i := len(extenderProfiles) - n; i < len(extenderProfiles); i += 1 {
+				// weight by success - error descending
+				extenderProfileIps[extenderProfiles[i]] = WeightedSelect(visitedExtenderIps)
+			}
+		}
+	} else {
+		ips := maps.Keys(self.extenderIpSecrets)
+		for _, extenderProfile := range extenderProfiles {
+			ip := ips[mathrand.Intn(len(ips))]
+			extenderConfig := &ExtenderConfig{
+				ExtenderProfile: extenderProfile,
+				Ip: ip,
+				Secret: self.extenderIpSecrets[ip],
+			}
+			extenderConfigs = append(extenderConfigs, extenderConfig)
+		}
+	}
+
+
+
+
+	expandedDialers := []*clientDialer{}
+	for _, extenderConfig := range extenderConfigs {
+		dialer := &net.Dialer{
+	        Timeout:   30 * time.Second,
+	        KeepAlive: 30 * time.Second,
+	    }
+	    dialTlsContext: NewExtenderDialTlsContext(
+		    dialer,
+		    extenderConfig,
+		    settings.tlsConfig,
+		)
+
+		dialer := &clientDialer{
+			dialTlsContext: dialTlsContext,
+			extenderProfile: extenderProfile,
+			extenderIp: extenderIp,
+		}
+		expandedDialers = append(expandedDialers, dialer)
+	}
+
+
+
+	self.dialers = append(self.dialers, extenderDialers...)
+	self.resolvedExtenderIps = append(self.resolvedExtenderIps, expandedExtenderIps...)
+
+
+
+
+
+
+	return expandedDialers, expandedExtenderIps
+}
+
